+++ LabeledIconButton/LabeledIconButton.js
@@ -18,13 +18,13 @@
 
 import kind from '@enact/core/kind';
 import Spottable from '@enact/spotlight/Spottable';
-import {IconButtonDecorator as UiIconButtonDecorator} from '@enact/ui/IconButton';
 import {LabeledIconBase as UiLabeledIconBase, LabeledIconDecorator as UiLabeledIconDecorator} from '@enact/ui/LabeledIcon';
 import Pure from '@enact/ui/internal/Pure';
 import PropTypes from 'prop-types';
 import compose from 'ramda/src/compose';
 
 import {IconButtonBase} from '../IconButton';
+import {IconButtonDecorator as UiIconButtonDecorator} from '../internal/IconButton';
 import Skinnable from '../Skinnable';
 
 import componentCss from './LabeledIconButton.module.less';