+++ SlotItem/SlotItem.js
@@ -23,12 +23,12 @@
 import kind from '@enact/core/kind';
 import Spottable from '@enact/spotlight/Spottable';
 import Pure from '@enact/ui/internal/Pure';
-import {SlotItemBase as UiSlotItemBase, SlotItemDecorator as UiSlotItemDecorator} from '@enact/ui/SlotItem';
 import {ItemDecorator as UiItemDecorator} from '@enact/ui/Item';
 import Toggleable from '@enact/ui/Toggleable';
 import PropTypes from 'prop-types';
 import compose from 'ramda/src/compose';
 
+import {SlotItemBase as UiSlotItemBase, SlotItemDecorator as UiSlotItemDecorator} from '../internal/SlotItem';
 import {ItemBase} from '../Item';
 import {MarqueeDecorator} from '../Marquee';
 import Skinnable from '../Skinnable';
@@ -40,7 +40,7 @@ import componentCss from './SlotItem.module.less';
  *
  * @class SlotItemBase
  * @memberof moonstone/SlotItem
- * @extends ui/SlotItem.SlotItemBase
+ * @extends moonstone/internal/SlotItem.SlotItemBase
  * @omit component
  * @mixes moonstone/Item.ItemBase
  * @ui