+++ samples/sampler/stories/qa/Scroller.js
@@ -1,6 +1,7 @@
 import Button from '@enact/moonstone/Button';
 import ExpandableList from '@enact/moonstone/ExpandableList';
 import Heading from '@enact/moonstone/Heading';
+import {ScrollableBase as UiScrollableBase} from '@enact/moonstone/internal/Scrollable';
 import Item from '@enact/moonstone/Item';
 import {ScrollableBase} from '@enact/moonstone/Scrollable';
 import Scroller from '@enact/moonstone/Scroller';
@@ -12,7 +13,6 @@ import ri from '@enact/ui/resolution';
 import Spotlight from '@enact/spotlight';
 import SpotlightContainerDecorator from '@enact/spotlight/SpotlightContainerDecorator';
 import UiScroller from '@enact/ui/Scroller';
-import {ScrollableBase as UiScrollableBase} from '@enact/ui/Scrollable';
 import {storiesOf} from '@storybook/react';
 import PropTypes from 'prop-types';
 import {useCallback, useEffect, useState} from 'react';