+++ samples/sampler/stories/qa/VirtualGridList.js
@@ -1,13 +1,14 @@
 import Button from '@enact/moonstone/Button';
 import ContextualPopupDecorator from '@enact/moonstone/ContextualPopupDecorator';
 import GridListImageItem from '@enact/moonstone/GridListImageItem';
+import {ScrollableBase as UiScrollableBase} from '@enact/moonstone/internal/Scrollable';
 import Item from '@enact/moonstone/Item';
 import {VirtualGridList, VirtualListBase} from '@enact/moonstone/VirtualList';
 import {action} from '@enact/storybook-utils/addons/actions';
 import {boolean, number, select} from '@enact/storybook-utils/addons/knobs';
 import {mergeComponentMetadata} from '@enact/storybook-utils';
 import ri from '@enact/ui/resolution';
-import {ScrollableBase as UiScrollableBase} from '@enact/ui/Scrollable';
+
 import {VirtualListBase as UiVirtualListBase} from '@enact/ui/VirtualList/VirtualListBase';
 import {storiesOf} from '@storybook/react';
 import PropTypes from 'prop-types';